打标工具/requirements.txt
requests>=2.31.0
Pillow>=10.0.0
tqdm>=4.65.0
PyMuPDF>=1.24.0
